Move Win32 -> POSIX error code conversion to runtime/win32.c
Use it in `caml_win32_rename`.
